削除のリファクタリング
